Update project files

+++ Assets/Scripts/Objects/RayTracingMesh.cs
@@ -45,20 +45,21 @@ namespace Objects {
         }
 
         public void UpdateMeshData() {
+            //Clear the current list of triangles
             Triangles.Clear();
             
+            //Get the vertices, normals, and indices of the mesh
             Mesh mesh = GetComponent<MeshFilter>().sharedMesh;
             Vector3[] vertices = mesh.vertices;
             Vector3[] normals = mesh.normals;
             int[] indices = mesh.triangles;
             
+            //Set the mesh name and the location to save the BVH data
             meshName = mesh.name;
             string filePath = Path.Combine(saveDir, meshName + "_BVH.json");
             
-            // Generate BVH
             // Check if the BVH file exists
-            if (File.Exists(filePath))
-            {
+            if (File.Exists(filePath)) {
                 try {
                     // Load the BVH file
                     LoadBVHFromFile(filePath);
@@ -66,14 +67,13 @@ namespace Objects {
                 catch (Exception ex) {
                     Debug.LogError($"Failed to load BVH: {ex.Message}");
                 }
-            }
-            else
-            {
+            } else {
                 // Generate a new BVH and save it to file
                 BVH = new BVH(vertices, indices, normals);
                 SaveBVHToFile(filePath);
             }
 
+            //Create a new stats object and save the data to it
             stats = new BvhStats {
                 MaxLeafDepth = int.MinValue, 
                 MinLeafDepth = int.MaxValue, 
@@ -82,16 +82,18 @@ namespace Objects {
                 LeafCount = BVH.LeafNodes.Count,
                 GenTime = BVH.genTime / 1000f
             };
-
+            
+            //Save the max and min leaf depth into stats
             foreach (int leafDepth in BVH.LeafNodesDepth) {
                 stats.MaxLeafDepth = Mathf.Max(stats.MaxLeafDepth, leafDepth);
                 stats.MinLeafDepth = Mathf.Min(stats.MinLeafDepth, leafDepth);
             }
 
+            //Save the mean leaf depth to stats
             stats.MeanLeafDepth = BVH.LeafNodesDepth.Sum() / BVH.LeafNodesDepth.Count;
             
+            //Save the max and min leaf tri count into stats
             int leafNodeTriCount = 0;
-
             foreach (CNode leaf in BVH.LeafNodes) {
                 stats.MaxLeafTri = Mathf.Max(stats.MaxLeafTri, leaf.TrianglesCount);
                 stats.MinLeafTri = Mathf.Min(stats.MinLeafTri, leaf.TrianglesCount);
@@ -99,8 +101,10 @@ namespace Objects {
                 leafNodeTriCount += leaf.TrianglesCount;
             }
 
-            stats.MeanLeafTri = leafNodeTriCount / BVH.LeafNodes.Count;
+            //Save the mean leaf tri count to stats
+            stats.MeanLeafTri = (float)leafNodeTriCount / BVH.LeafNodes.Count;
 
+            //Create a new triangle and add it to the triangle array
             for (int i = 0; i < indices.Length; i += 3) {
                 MeshTriangle meshTriangle = new();
 
@@ -121,12 +125,15 @@ namespace Objects {
             }
         }
         
-        private void SaveBVHToFile(string filePath)
-        {
+        private void SaveBVHToFile(string filePath) {
+            //If the file directory does not exist, create it
             if (!Directory.Exists(saveDir)) Directory.CreateDirectory(saveDir);
 
+            //Open the file into a binary writer
             using BinaryWriter writer = new(File.Open(filePath, FileMode.Create));
-            writer.Write(BVH.genTime); // Save generation time
+            
+            //Save generation time
+            writer.Write(BVH.genTime); 
 
             // Save Root Node
             WriteNode(writer, BVH.Root);
@@ -149,18 +156,23 @@ namespace Objects {
         }
 
         private void LoadBVHFromFile(string filePath) {
+            //Create a new BVH to load data into
             BVH = new BVH();
             
+            //Wait if the file is locked to avoid errors
             while (IsFileLocked(new FileInfo(filePath)))
-                // Optionally, add a small delay to avoid a busy wait
                 System.Threading.Thread.Sleep(100);
 
+            //Open the file into the reader
             using BinaryReader reader = new(File.Open(filePath, FileMode.Open));
-            BVH.genTime = reader.ReadSingle(); // Read generation time
+            
+            // Read generation time
+            BVH.genTime = reader.ReadSingle(); 
 
             // Read Root Node
             BVH.Root = ReadNode(reader);
 
+            //Create new lists to load data into
             BVH.AllNodes = new List<Node>();
             BVH.LeafNodes = new List<CNode>();
             BVH.LeafNodesDepth = new List<int>();
@@ -188,27 +200,29 @@ namespace Objects {
         
         private bool IsFileLocked(FileInfo file)
         {
+            //Create a blank stream
             FileStream stream = null;
 
-            try
-            {
+            try {
+                //Attempt to open the file into the stream
                 stream = file.Open(FileMode.Open, FileAccess.Read, FileShare.None);
             }
-            catch (IOException)
-            {
-                // File is locked by another process
+            catch (IOException) {
+                //Return true if the file is locked by another process
                 return true;
             }
-            finally
-            {
+            finally {
+                //Close the stream if it has been opened
                 stream?.Close();
             }
 
+            //Return false if the file was not locked
             return false;
         }
 
         private void WriteNode(BinaryWriter writer, Node node)
         {
+            //Write each individual data of a node
             WriteVector3(writer, node.BoundsMin);
             WriteVector3(writer, node.BoundsMax);
             writer.Write(node.TriangleIndex);
@@ -218,16 +232,19 @@ namespace Objects {
 
         private Node ReadNode(BinaryReader reader)
         {
+            //Read each individual chunk of data for a node
             Vector3 boundsMin = ReadVector3(reader);
             Vector3 boundsMax = ReadVector3(reader);
             int triangleIndex = reader.ReadInt32();
             int childIndex = reader.ReadInt32();
             int trianglesCount = reader.ReadInt32();
-            return new Node { BoundsMin = boundsMin, BoundsMax = boundsMax, TriangleIndex = triangleIndex, ChildIndex = childIndex, TrianglesCount = trianglesCount };
+            return new Node { BoundsMin = boundsMin, BoundsMax = boundsMax, TriangleIndex = triangleIndex, ChildIndex = childIndex, 
+                TrianglesCount = trianglesCount };
         }
 
         private void WriteTriangle(BinaryWriter writer, MeshTriangle meshTriangle)
         {
+            //Write each data point of the triangle
             WriteVector3(writer, meshTriangle.posA);
             WriteVector3(writer, meshTriangle.posB);
             WriteVector3(writer, meshTriangle.posC);
@@ -238,6 +255,7 @@ namespace Objects {
 
         private MeshTriangle ReadTriangle(BinaryReader reader)
         {
+            //Read all the data for a triangle
             Vector3 posA = ReadVector3(reader);
             Vector3 posB = ReadVector3(reader);
             Vector3 posC = ReadVector3(reader);
@@ -249,6 +267,7 @@ namespace Objects {
 
         private void WriteCNode(BinaryWriter writer, CNode cnode)
         {
+            //Write each data of a CNode
             WriteBoundingBox(writer, cnode.Bounds);
             writer.Write(cnode.TriangleIndex);
             writer.Write(cnode.ChildIndex);
@@ -257,6 +276,7 @@ namespace Objects {
 
         private CNode ReadCNode(BinaryReader reader)
         {
+            //Read all the data of a CNode
             CBoundingBox bounds = ReadBoundingBox(reader);
             int triangleIndex = reader.ReadInt32();
             int childIndex = reader.ReadInt32();
@@ -266,6 +286,7 @@ namespace Objects {
 
         private static void WriteVector3(BinaryWriter writer, Vector3 vector)
         {
+            //Write each value for the Vector3
             writer.Write(vector.x);
             writer.Write(vector.y);
             writer.Write(vector.z);
@@ -273,6 +294,7 @@ namespace Objects {
 
         private static Vector3 ReadVector3(BinaryReader reader)
         {
+            //Read each value for the Vector3
             float x = reader.ReadSingle();
             float y = reader.ReadSingle();
             float z = reader.ReadSingle();
@@ -281,12 +303,14 @@ namespace Objects {
 
         private static void WriteBoundingBox(BinaryWriter writer, CBoundingBox bounds)
         {
+            //Write each data of a bounding box
             WriteVector3(writer, bounds.Min);
             WriteVector3(writer, bounds.Max);
         }
 
         private static CBoundingBox ReadBoundingBox(BinaryReader reader)
         {
+            //Read the data for a bounding box
             Vector3 min = ReadVector3(reader);
             Vector3 max = ReadVector3(reader);
             return new CBoundingBox { Min = min, Max = max };
